+++ CHANGELOG.md
@@ -2,6 +2,12 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 ## [Unreleased]
 
+# v22.10.14
+
+### Enhancements
+
+* Separate update intervals for OEIS files (7 days) and programs repository (1 day)
+
 # v22.10.13
 
 ### Features